Moved IGreeter to Services folder, added a constructor to the home controller that builds our list of in-memory restaurants and will return that list on the Index page, and added a scoped service for our in-memory restaurant data.
